Consolidated advanced aerospace and small craft/dropship critical allocation into a single class.
