+++ src/megameklab/com/ui/Dropship/tabs/DropshipBuildTab.java
@@ -28,8 +28,7 @@ import megamek.common.Mounted;
 import megameklab.com.ui.EntitySource;
 import megameklab.com.ui.Aero.tabs.EquipmentTab;
 import megameklab.com.ui.Dropship.views.AerospaceBuildView;
-import megameklab.com.ui.Dropship.views.DropshipCriticalView;
-import megameklab.com.ui.advaero.AdvancedAeroCriticalView;
+import megameklab.com.ui.Dropship.views.LargeCraftCritcalView;
 import megameklab.com.util.ITab;
 import megameklab.com.util.RefreshListener;
 import megameklab.com.util.UnitUtil;
@@ -48,7 +47,7 @@ public class DropshipBuildTab extends ITab implements ActionListener {
     private static final long serialVersionUID = -6625026210117558378L;
     
     private RefreshListener refresh = null;
-    private JPanel critView = null;
+    private LargeCraftCritcalView critView = null;
     private AerospaceBuildView buildView = null;
     private JPanel buttonPanel = new JPanel();
     private JPanel mainPanel = new JPanel();
@@ -66,13 +65,8 @@ public class DropshipBuildTab extends ITab implements ActionListener {
         GridBagConstraints gbc = new GridBagConstraints();
 
         buildView = new AerospaceBuildView(eSource,refresh);
-        if (eSource.getEntity().hasETypeFlag(Entity.ETYPE_JUMPSHIP)) {
-            critView = new AdvancedAeroCriticalView(eSource, refresh);
-            ((AdvancedAeroCriticalView) critView).addAllocationListeners(buildView);
-        } else {
-            critView = new DropshipCriticalView(eSource, refresh);
-            ((DropshipCriticalView) critView).addAllocationListeners(buildView);
-        }
+        critView = new LargeCraftCritcalView(eSource, refresh);
+        critView.addAllocationListeners(buildView);
 
         resetButton.setMnemonic('R');
         resetButton.setActionCommand(RESETCOMMAND);        
@@ -98,11 +92,7 @@ public class DropshipBuildTab extends ITab implements ActionListener {
 
     public void refresh() {
         removeAllActionListeners();
-        if (eSource.getEntity().hasETypeFlag(Entity.ETYPE_JUMPSHIP)) {
-            ((AdvancedAeroCriticalView) critView).refresh();
-        } else {
-            ((DropshipCriticalView) critView).refresh();
-        }
+        critView.refresh();
         critView.validate();
         buildView.refresh();
         addAllActionListeners();
@@ -135,11 +125,7 @@ public class DropshipBuildTab extends ITab implements ActionListener {
 
     public void addRefreshedListener(RefreshListener l) {
         refresh = l;
-        if (eSource.getEntity().hasETypeFlag(Entity.ETYPE_JUMPSHIP)) {
-            ((AdvancedAeroCriticalView) critView).updateRefresh(refresh);
-        } else {
-            ((DropshipCriticalView) critView).updateRefresh(refresh);
-        }
+        critView.updateRefresh(refresh);
         buildView.addRefreshedListener(refresh);
     }
