+++ src/megameklab/resources/Views.properties
@@ -265,8 +265,7 @@ BAEnhancementView.chkMyomerBooster.text=Myomer Booster
 BAEnhancementView.chkMyomerBooster.tooltip=Increases ground MP. Clan tech base only.
 
 DropshipCriticalView.aerodyneArcs.values=Nose,Left Wing,Right Wing,Aft
-DropshipCriticalView.spheroidArcs.values=Nose,Forward Left,Foreward Right,Aft,Aft Left,Aft Right
-AdvancedAeroCriticalView.arcs.values=Nose,Forward Left,Foreward Right,Aft,Aft Left,Aft Right,Left Broadside,Right Broadside
+DropshipCriticalView.spheroidArcs.values=Nose,Forward Left,Forward Right,Aft,Aft Left,Aft Right,Left Broadside,Right Broadside
 DropshipCriticalView.lblSlotCount.text=Slots used:
 DropshipCriticalView.lblSlotCount.tooltip=The number of slots in this arc occupied by weapons
 DropshipCriticalView.lblMaxSlots.format=Weight-free allowed: %d
